+++ modules/admin/src/views/aws/info/index.php
@@ -1,9 +1,9 @@
-<h1>Detail</h1>
-<table class="striped">
-<?php foreach ($data as $key => $value): ?>
-<tr>
-    <td style="width:20%"><?= $key; ?></td>
-    <td style="width:80%"><?= $value; ?></td>
-</tr>
-<?php endforeach; ?>
-</table>
+<?php
+use yii\widgets\DetailView;
+use yii\base\Widget;
+?>
+<?= DetailView::widget([
+    'model' => $model,
+    'attributes' => $attributes,
+    'options' => ['class' => 'striped highlight bordered responsive-table'],
+]); ?>